// src/api/model/loadOptionsDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { GroupingInfo } from "./groupingInfo";
import type { SortingInfo } from "./sortingInfo";
import type { SummaryInfo } from "./summaryInfo";

export interface LoadOptionsDto {
	requireTotalCount?: boolean;
	requireGroupCount?: boolean;
	isCountQuery?: boolean;
	isSummaryQuery?: boolean;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	skip?: number | string;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	take?: number | string;
	/** @nullable */
	sort?: SortingInfo[] | null;
	/** @nullable */
	group?: GroupingInfo[] | null;
	/** @nullable */
	filter?: unknown[] | null;
	/** @nullable */
	totalSummary?: SummaryInfo[] | null;
	/** @nullable */
	groupSummary?: SummaryInfo[] | null;
	/** @nullable */
	select?: string[] | null;
	/** @nullable */
	preSelect?: string[] | null;
	/** @nullable */
	remoteSelect?: boolean | null;
	/** @nullable */
	remoteGrouping?: boolean | null;
	/** @nullable */
	expandLinqSumType?: boolean | null;
	/** @nullable */
	primaryKey?: string[] | null;
	/** @nullable */
	defaultSort?: string | null;
	/** @nullable */
	stringToLower?: boolean | null;
	/** @nullable */
	paginateViaPrimaryKey?: boolean | null;
	/** @nullable */
	sortByPrimaryKey?: boolean | null;
	allowAsyncOverSync?: boolean;
}